refactor: cvedb structure
