+++ cve_bin_tool/cli.py
@@ -53,6 +53,7 @@ from cve_bin_tool.merge import MergeReports
 from cve_bin_tool.output_engine import OutputEngine
 from cve_bin_tool.package_list_parser import PackageListParser
 from cve_bin_tool.sbom_manager import SBOMManager
+from cve_bin_tool.sources import curl, nvd
 from cve_bin_tool.util import ProductInfo
 from cve_bin_tool.version import VERSION
 from cve_bin_tool.version_scanner import VersionScanner
@@ -449,14 +450,17 @@ def main(argv=None):
             "Use -F --filter only when you want to filter out intermediate reports on the basis of tag"
         )
 
-    # Database update related settings
-    # Connect to the database
-    cvedb_orig = CVEDB(
-        version_check=not version_check,
-        error_mode=error_mode,
+    nvd_source = nvd.NVD(
         nvd_type=args["nvd"],
         incremental_update=True if db_update == "latest" and args["nvd"] else False,
         nvd_api_key=args["nvd_api_key"],
+    )
+    # Database update related settings
+    # Connect to the database
+    cvedb_orig = CVEDB(
+        sources=[nvd_source, curl.Curl()],
+        version_check=not version_check,
+        error_mode=error_mode,
     )
 
     # if OLD_CACHE_DIR (from cvedb.py) exists, print warning
@@ -489,7 +493,7 @@ def main(argv=None):
         if args["nvd"] == "json":
             LOGGER.warning("Not verifying CVE DB cache")
             cvedb_orig.get_db_update_date()
-            if not cvedb_orig.nvd_years():
+            if not nvd_source.nvd_years():
                 with ErrorHandler(mode=error_mode, logger=LOGGER):
                     raise EmptyCache(cvedb_orig.cachedir)
