+++ cve_bin_tool/cvedb.py
@@ -1,41 +1,26 @@
-# Copyright (C) 2021 Intel Corporation
+# Copyright (C) 2022 Intel Corporation
 # SPDX-License-Identifier: GPL-3.0-or-later
 
 """
-Retrieval access and caching of NIST CVE database
+Handling CVE database
 """
 from __future__ import annotations
 
 import asyncio
 import datetime
-import glob
-import gzip
-import hashlib
-import json
 import logging
 import os
-import re
 import shutil
 import sqlite3
 from typing import Any
 
-import aiohttp
 import requests
-from bs4 import BeautifulSoup
 from rich.progress import track
 
-from cve_bin_tool.async_utils import FileIO, GzipFile, RateLimiter, run_coroutine
-from cve_bin_tool.error_handler import (
-    AttemptedToWriteOutsideCachedir,
-    CVEDataForCurlVersionNotInCache,
-    CVEDataForYearNotInCache,
-    ErrorHandler,
-    ErrorMode,
-    NVDRateLimit,
-    SHAMismatch,
-)
+from cve_bin_tool.async_utils import run_coroutine
+from cve_bin_tool.error_handler import ErrorMode
 from cve_bin_tool.log import LOGGER
-from cve_bin_tool.nvd_api import NVD_API
+from cve_bin_tool.sources import curl, nvd
 from cve_bin_tool.version import check_latest_version
 
 logging.basicConfig(level=logging.DEBUG)
@@ -47,42 +32,33 @@ DISK_LOCATION_BACKUP = os.path.join(
 )
 DBNAME = "cve.db"
 OLD_CACHE_DIR = os.path.join(os.path.expanduser("~"), ".cache", "cvedb")
-NVD_FILENAME_TEMPLATE = "nvdcve-1.1-{}.json.gz"
 
 
 class CVEDB:
     """
-    Downloads NVD data in json form and stores it on disk in a cache.
+    Retrieves CVE data from sources and handles CVE Database
     """
 
     CACHEDIR = DISK_LOCATION_DEFAULT
     BACKUPCACHEDIR = DISK_LOCATION_BACKUP
-    FEED = "https://nvd.nist.gov/vuln/data-feeds"
     LOGGER = LOGGER.getChild("CVEDB")
-    NVDCVE_FILENAME_TEMPLATE = NVD_FILENAME_TEMPLATE
-    CURL_CVE_FILENAME_TEMPLATE = "curlcve-{}.json"
-    META_LINK = "https://nvd.nist.gov"
-    META_REGEX = re.compile(r"\/feeds\/json\/.*-[0-9]*\.[0-9]*-[0-9]*\.meta")
-    RANGE_UNSET = ""
+    SOURCES = [nvd.NVD(), curl.Curl()]
 
     def __init__(
         self,
-        feed: str | None = None,
+        sources=None,
         cachedir: str | None = None,
         backup_cachedir: str | None = None,
         version_check: bool = True,
-        session: RateLimiter | None = None,
         error_mode: ErrorMode = ErrorMode.TruncTrace,
-        nvd_type: str = "json",
-        incremental_update: bool = False,
-        nvd_api_key: str = "",
     ):
-        self.feed = feed if feed is not None else self.FEED
+        self.sources = sources if sources is not None else self.SOURCES
         self.cachedir = cachedir if cachedir is not None else self.CACHEDIR
         self.backup_cachedir = (
             backup_cachedir if backup_cachedir is not None else self.BACKUPCACHEDIR
         )
         self.error_mode = error_mode
+
         # Will be true if refresh was successful
         self.was_updated = False
 
@@ -92,16 +68,13 @@ class CVEDB:
         # set up the db if needed
         self.dbpath = os.path.join(self.cachedir, DBNAME)
         self.connection: sqlite3.Connection | None = None
-        self.session = session
+
+        self.data = []
         self.cve_count = -1
-        self.nvd_type = nvd_type
-        self.incremental_update = incremental_update
         self.all_cve_entries: list[dict[str, Any]] | None = None
+
         self.exploits_list = []
         self.exploit_count = 0
-
-        # store the nvd api key for use later
-        self.nvd_api_key = nvd_api_key
 
         if not os.path.exists(self.dbpath):
             self.rollback_cache_backup()
@@ -111,27 +84,6 @@ class CVEDB:
             # Force update
             self.check_cve_entries()
         return self.cve_count
-
-    async def nist_fetch_using_api(self) -> list:
-        """Fetch using NVD's CVE API (as opposed to NVD's JSON Vulnerability Feeds)"""
-        nvd_api = NVD_API(
-            logger=self.LOGGER,
-            error_mode=self.error_mode,
-            incremental_update=self.incremental_update,
-            api_key=self.nvd_api_key,
-        )
-        if self.incremental_update:
-            await nvd_api.get_nvd_params(
-                time_of_last_update=datetime.datetime.fromtimestamp(
-                    self.get_db_update_date()
-                )
-            )
-        else:
-            await nvd_api.get_nvd_params()
-        await nvd_api.get()
-        await nvd_api.session.close()
-        nvd_api.session = None
-        return nvd_api.all_cve_entries
 
     def check_db_exists(self) -> bool:
         return os.path.isfile(self.dbpath)
@@ -143,133 +95,6 @@ class CVEDB:
         )
         return os.path.getmtime(self.dbpath)
 
-    async def getmeta(
-        self, session: RateLimiter, meta_url: str
-    ) -> tuple[str, dict[str, str]]:
-        async with await session.get(meta_url) as response:
-            response.raise_for_status()
-            return (
-                meta_url.replace(".meta", ".json.gz"),
-                dict(
-                    [
-                        line.split(":", maxsplit=1)
-                        for line in (await response.text()).splitlines()
-                        if ":" in line
-                    ]
-                ),
-            )
-
-    async def nist_scrape(self, session: RateLimiter):
-        async with await session.get(self.feed) as response:
-            response.raise_for_status()
-            page = await response.text()
-            json_meta_links = self.META_REGEX.findall(page)
-            return dict(
-                await asyncio.gather(
-                    *(
-                        self.getmeta(session, f"{self.META_LINK}{meta_url}")
-                        for meta_url in json_meta_links
-                    )
-                )
-            )
-
-    async def cache_update(
-        self,
-        session: RateLimiter,
-        url: str,
-        sha: str,
-        chunk_size: int = 16 * 1024,
-    ) -> None:
-        """
-        Update the cache for a single year of NVD data.
-        """
-        filename = url.split("/")[-1]
-        # Ensure we only write to files within the cachedir
-        filepath = os.path.abspath(os.path.join(self.cachedir, filename))
-        if not filepath.startswith(os.path.abspath(self.cachedir)):
-            with ErrorHandler(mode=self.error_mode, logger=self.LOGGER):
-                raise AttemptedToWriteOutsideCachedir(filepath)
-        # Validate the contents of the cached file
-        if os.path.isfile(filepath):
-            # Validate the sha and write out
-            sha = sha.upper()
-            calculate = hashlib.sha256()
-            async with GzipFile(filepath, "rb") as f:
-                chunk = await f.read(chunk_size)
-                while chunk:
-                    calculate.update(chunk)
-                    chunk = await f.read(chunk_size)
-            # Validate the sha and exit if it is correct, otherwise update
-            gotsha = calculate.hexdigest().upper()
-            if gotsha != sha:
-                os.unlink(filepath)
-                self.LOGGER.debug(
-                    f"SHA mismatch for {filename} (have: {gotsha}, want: {sha})"
-                )
-            else:
-                self.LOGGER.debug(f"Correct SHA for {filename}")
-                return
-        self.LOGGER.debug(f"Updating CVE cache for {filename}")
-        async with await session.get(url) as response:
-            # Raise better error message on ratelimit by NVD
-            if response.status == 403:
-                with ErrorHandler(mode=self.error_mode, logger=self.LOGGER):
-                    raise NVDRateLimit(
-                        f"{url} : download failed, you may have been rate limited."
-                    )
-            # Raise for all other 4xx errors
-            response.raise_for_status()
-            gzip_data = await response.read()
-        json_data = gzip.decompress(gzip_data)
-        gotsha = hashlib.sha256(json_data).hexdigest().upper()
-        async with FileIO(filepath, "wb") as filepath_handle:
-            await filepath_handle.write(gzip_data)
-        # Raise error if there was an issue with the sha
-        if gotsha != sha:
-            # Remove the file if there was an issue
-            # exit(100)
-            os.unlink(filepath)
-            with ErrorHandler(mode=self.error_mode, logger=self.LOGGER):
-                raise SHAMismatch(f"{url} (have: {gotsha}, want: {sha})")
-
-    @staticmethod
-    async def get_curl_versions(session: RateLimiter) -> list[str]:
-        regex = re.compile(r"vuln-(\d+.\d+.\d+)\.html")
-        async with await session.get(
-            "https://curl.haxx.se/docs/vulnerabilities.html"
-        ) as response:
-            response.raise_for_status()
-            html = await response.text()
-        matches = regex.finditer(html)
-        return [match.group(1) for match in matches]
-
-    async def download_curl_version(self, session: RateLimiter, version: str) -> None:
-        async with await session.get(
-            f"https://curl.haxx.se/docs/vuln-{version}.html"
-        ) as response:
-            response.raise_for_status()
-            html = await response.text()
-        soup = BeautifulSoup(html, "html.parser")
-        table = soup.find("table")
-        if not table:
-            return
-        headers = table.find_all("th")
-        headers = list(map(lambda x: x.text.strip().lower(), headers))
-        self.LOGGER.debug(headers)
-        rows = table.find_all("tr")
-        json_data = []
-        for row in rows:
-            cols = row.find_all("td")
-            values = (ele.text.strip() for ele in cols)
-            data = dict(zip(headers, values))
-            if data:
-                json_data.append(data)
-        filepath = os.path.abspath(
-            os.path.join(self.cachedir, f"curlcve-{version}.json")
-        )
-        async with FileIO(filepath, "w") as f:
-            await f.write(json.dumps(json_data, indent=4))
-
     async def refresh(self) -> None:
         """Refresh the cve database and check for new version."""
         # refresh the database
@@ -279,65 +104,8 @@ class CVEDB:
         # check for the latest version
         if self.version_check:
             check_latest_version()
-        if not self.session:
-            connector = aiohttp.TCPConnector(limit_per_host=19)
-            self.session = RateLimiter(
-                aiohttp.ClientSession(connector=connector, trust_env=True)
-            )
 
-        tasks = []
-        if self.nvd_type == "api":
-            self.LOGGER.info("[Using NVD API]")
-            self.all_cve_entries, curl_metadata = await asyncio.gather(
-                self.nist_fetch_using_api(),
-                self.get_curl_versions(self.session),
-            )
-
-        else:
-            self.LOGGER.info("Downloading CVE data...")
-            nvd_metadata, curl_metadata = await asyncio.gather(
-                self.nist_scrape(self.session),
-                self.get_curl_versions(self.session),
-            )
-
-            tasks = [
-                self.cache_update(self.session, url, meta["sha256"])
-                for url, meta in nvd_metadata.items()
-                if meta is not None
-            ]
-
-        # We use gather to create a single task from a set of tasks
-        # which download CVEs for each version of curl. Otherwise
-        # the progress bar would show that we are closer to
-        # completion than we think, because lots of curl CVEs (for
-        # each version) have been downloaded
-        tasks.append(
-            asyncio.gather(
-                *(
-                    asyncio.ensure_future(
-                        self.download_curl_version(self.session, version)
-                    )
-                    for version in curl_metadata
-                )
-            )
-        )
-        total_tasks = len(tasks)
-
-        # error_mode.value will only be greater than 1 if quiet mode.
-        if self.error_mode.value > 1:
-            iter_tasks = track(
-                asyncio.as_completed(tasks),
-                description="Downloading CVEs...",
-                total=total_tasks,
-            )
-        else:
-            iter_tasks = asyncio.as_completed(tasks)
-
-        for task in iter_tasks:
-            await task
-        self.was_updated = True
-        await self.session.close()
-        self.session = None
+        await self.get_data()
 
     def refresh_cache_and_update_db(self) -> None:
         self.LOGGER.debug("Updating CVE data. This will take a few minutes.")
@@ -439,6 +207,16 @@ class CVEDB:
             self.clear_cached_data()
         self.connection.commit()
 
+    async def get_data(self):
+        tasks = []
+
+        for source in self.sources:
+            if source is not None:
+                tasks.append(source.get_cve_data())
+
+        for r in await asyncio.gather(*tasks):
+            self.data.append(r)
+
     def populate_db(self) -> None:
         """Function that populates the database from the JSON.
 
@@ -446,9 +224,27 @@ class CVEDB:
         in single nodes.  This isn't *required* by the json so may not be true everywhere.  If that's the case,
         we'll need a better parser to match those together.
         """
+
         self.db_open()
         cursor = self.connection.cursor()
 
+        for severity_data, affected_data in self.data:
+            if severity_data is not None:
+                self.populate_severity(
+                    severity_data,
+                    cursor,
+                )
+            if affected_data is not None:
+                self.populate_affected(
+                    affected_data,
+                    cursor,
+                )
+
+        self.connection.commit()
+
+        self.db_close()
+
+    def populate_severity(self, severity_data, cursor):
         insert_severity = """
         INSERT or REPLACE INTO cve_severity(
             CVE_number,
@@ -460,106 +256,10 @@ class CVEDB:
         )
         VALUES (?, ?, ?, ?, ?, ?)
         """
-        insert_cve_range = """
-        INSERT or REPLACE INTO cve_range(
-            cve_number,
-            vendor,
-            product,
-            version,
-            versionStartIncluding,
-            versionStartExcluding,
-            versionEndIncluding,
-            versionEndExcluding
-        ) VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-        """
+
         del_cve_range = "DELETE from cve_range where CVE_number=?"
-        if self.nvd_type == "api":
-            self.LOGGER.debug("Populating database from NVD API")
-            self.populate_cve_list(
-                self.all_cve_entries,
-                cursor,
-                insert_severity,
-                insert_cve_range,
-                del_cve_range,
-            )
-        else:
-            # error_mode.value will only be greater than 1 if quiet mode.
-            if self.error_mode.value > 1:
-                years = self.nvd_years()
-            else:
-                years = track(self.nvd_years(), description="Updating CVEs from NVD...")
 
-            for year in years:
-                cve_data = self.load_nvd_year(year)
-                self.LOGGER.debug(
-                    f'Time = {datetime.datetime.today().strftime("%H:%M:%S")}'
-                )
-                self.populate_cve_list(
-                    cve_data["CVE_Items"],
-                    cursor,
-                    insert_severity,
-                    insert_cve_range,
-                    del_cve_range,
-                )
-        self.connection.commit()
-
-        # supplemental data gets added here
-        self.supplement_curl()
-
-        self.db_close()
-
-    def populate_cve_list(
-        self,
-        all_cve_list: list[dict[str, Any]],
-        cursor: sqlite3.Cursor,
-        insert_severity: str,
-        insert_cve_range: str,
-        del_cve_range: str,
-    ):
-        """Populate the cve_list with the CVEs from the JSON"""
-        for cve_item in all_cve_list:
-            # the information we want:
-            # CVE ID, Severity, Score ->
-            # affected {Vendor(s), Product(s), Version(s)}
-            cve = {
-                "ID": cve_item["cve"]["CVE_data_meta"]["ID"],
-                "description": cve_item["cve"]["description"]["description_data"][0][
-                    "value"
-                ],
-                "severity": "unknown",
-                "score": "unknown",
-                "CVSS_version": "unknown",
-                "CVSS_vector": "unknown",
-            }
-            if cve["description"].startswith("** REJECT **"):
-                # Skip this CVE if it's marked as 'REJECT'
-                continue
-
-            # Get CVSSv3 or CVSSv2 score for output.
-            # Details are left as an exercise to the user.
-            if "baseMetricV3" in cve_item["impact"]:
-                cve["severity"] = cve_item["impact"]["baseMetricV3"]["cvssV3"][
-                    "baseSeverity"
-                ]
-                cve["score"] = cve_item["impact"]["baseMetricV3"]["cvssV3"]["baseScore"]
-                cve["CVSS_vector"] = cve_item["impact"]["baseMetricV3"]["cvssV3"][
-                    "vectorString"
-                ]
-                cve["CVSS_version"] = 3
-            elif "baseMetricV2" in cve_item["impact"]:
-                cve["severity"] = cve_item["impact"]["baseMetricV2"]["severity"]
-                cve["score"] = cve_item["impact"]["baseMetricV2"]["cvssV2"]["baseScore"]
-                cve["CVSS_vector"] = cve_item["impact"]["baseMetricV2"]["cvssV2"][
-                    "vectorString"
-                ]
-                cve["CVSS_version"] = 2
-
-            # self.LOGGER.debug(
-            #    "Severity: {} ({}) v{}".format(
-            #        CVE["severity"], CVE["score"], CVE["CVSS_version"]
-            #    )
-            # )
-
+        for cve in severity_data:
             cursor.execute(
                 insert_severity,
                 [
@@ -575,72 +275,8 @@ class CVEDB:
             # Delete any old range entries for this CVE_number
             cursor.execute(del_cve_range, (cve["ID"],))
 
-            # walk the nodes with version data
-            # return list of versions
-            affects_list = []
-            if "configurations" in cve_item:
-                for node in cve_item["configurations"]["nodes"]:
-                    # self.LOGGER.debug("NODE: {}".format(node))
-                    affects_list.extend(self.parse_node(node))
-                    if "children" in node:
-                        for child in node["children"]:
-                            affects_list.extend(self.parse_node(child))
-            # self.LOGGER.debug("Affects: {}".format(affects_list))
-            cursor.executemany(
-                insert_cve_range,
-                [
-                    (
-                        cve["ID"],
-                        affected["vendor"],
-                        affected["product"],
-                        affected["version"],
-                        affected["versionStartIncluding"],
-                        affected["versionStartExcluding"],
-                        affected["versionEndIncluding"],
-                        affected["versionEndExcluding"],
-                    )
-                    for affected in affects_list
-                ],
-            )
+    def populate_affected(self, affected_data, cursor):
 
-    def parse_node(self, node: dict[str, Any]) -> list[dict[str, str]]:
-        affects_list = []
-        if "cpe_match" in node:
-            for cpe_match in node["cpe_match"]:
-                # self.LOGGER.debug(cpe_match["cpe23Uri"])
-                cpe_split = cpe_match["cpe23Uri"].split(":")
-                affects = {
-                    "vendor": cpe_split[3],
-                    "product": cpe_split[4],
-                    "version": cpe_split[5],
-                }
-
-                # self.LOGGER.debug(
-                #    "Vendor: {} Product: {} Version: {}".format(
-                #        affects["vendor"], affects["product"], affects["version"]
-                #    )
-                # )
-                # if we have a range (e.g. version is *) fill it out, and put blanks where needed
-                range_fields = [
-                    "versionStartIncluding",
-                    "versionStartExcluding",
-                    "versionEndIncluding",
-                    "versionEndExcluding",
-                ]
-                for field in range_fields:
-                    if field in cpe_match:
-                        affects[field] = cpe_match[field]
-                    else:
-                        affects[field] = self.RANGE_UNSET
-
-                affects_list.append(affects)
-        return affects_list
-
-    def supplement_curl(self) -> None:
-        """
-        Get additional CVE data directly from the curl website amd add it to the cvedb
-        """
-        self.db_open()
         insert_cve_range = """
         INSERT or REPLACE INTO cve_range(
             cve_number,
@@ -653,87 +289,23 @@ class CVEDB:
             versionEndExcluding
         ) VALUES (?, ?, ?, ?, ?, ?, ?, ?)
         """
-        cursor = self.connection.cursor()
-        # No need to track this. It is very fast!
-        for version in self.curl_versions():
-            cve_list = self.load_curl_version(version)
-            # for cve in cve_list:
-            cursor.executemany(
-                insert_cve_range,
-                [
-                    (
-                        cve["cve"],
-                        "haxx",
-                        "curl",
-                        version,
-                        cve["from version"],
-                        "",
-                        cve["to and including"],
-                        "",
-                    )
-                    for cve in cve_list
-                ],
-            )
-            self.connection.commit()
 
-    def load_nvd_year(self, year: int) -> dict[str, Any]:
-        """
-        Return the dict of CVE data for the given year.
-        """
-        filename = os.path.join(
-            self.cachedir, self.NVDCVE_FILENAME_TEMPLATE.format(year)
+        cursor.executemany(
+            insert_cve_range,
+            [
+                (
+                    affected["cve_id"],
+                    affected["vendor"],
+                    affected["product"],
+                    affected["version"],
+                    affected["versionStartIncluding"],
+                    affected["versionStartExcluding"],
+                    affected["versionEndIncluding"],
+                    affected["versionEndExcluding"],
+                )
+                for affected in affected_data
+            ],
         )
-        # Check if file exists
-        if not os.path.isfile(filename):
-            with ErrorHandler(mode=self.error_mode, logger=self.LOGGER):
-                raise CVEDataForYearNotInCache(year)
-        # Open the file and load the JSON data, log the number of CVEs loaded
-        with gzip.open(filename, "rb") as fileobj:
-            cves_for_year = json.load(fileobj)
-            self.LOGGER.debug(
-                f'Year {year} has {len(cves_for_year["CVE_Items"])} CVEs in dataset'
-            )
-            return cves_for_year
-
-    def nvd_years(self) -> list[int]:
-        """
-        Return the years we have NVD data for.
-        """
-        return sorted(
-            int(filename.split(".")[-3].split("-")[-1])
-            for filename in glob.glob(
-                os.path.join(self.cachedir, "nvdcve-1.1-*.json.gz")
-            )
-        )
-
-    def load_curl_version(self, version: str) -> dict[str, Any]:
-        """
-        Return the dict of CVE data for the given curl version.
-        """
-        filename = os.path.join(
-            self.cachedir, self.CURL_CVE_FILENAME_TEMPLATE.format(version)
-        )
-        # Check if file exists
-        if not os.path.isfile(filename):
-            with ErrorHandler(mode=self.error_mode, logger=self.LOGGER):
-                raise CVEDataForCurlVersionNotInCache(version)
-        # Open the file and load the JSON data, log the number of CVEs loaded
-        with open(filename, "rb") as fileobj:
-            cves_for_version = json.load(fileobj)
-            self.LOGGER.debug(
-                f"Curl Version {version} has {len(cves_for_version)} CVEs in dataset"
-            )
-            return cves_for_version
-
-    def curl_versions(self) -> list[str]:
-        """
-        Return the versions we have Curl data for.
-        """
-        regex = re.compile(r"curlcve-(\d+.\d+.\d).json")
-        return [
-            regex.search(filename).group(1)
-            for filename in glob.glob(os.path.join(self.cachedir, "curlcve-*.json"))
-        ]
 
     def clear_cached_data(self) -> None:
         self.create_cache_backup()
