+++ cve_bin_tool/sources/__init__.py
@@ -1,0 +1,17 @@
+import os
+from abc import ABC, abstractmethod
+
+# database defaults
+DISK_LOCATION_DEFAULT = os.path.join(os.path.expanduser("~"), ".cache", "cve-bin-tool")
+DISK_LOCATION_BACKUP = os.path.join(
+    os.path.expanduser("~"), ".cache", "cve-bin-tool-backup"
+)
+DBNAME = "cve.db"
+OLD_CACHE_DIR = os.path.join(os.path.expanduser("~"), ".cache", "cvedb")
+NVD_FILENAME_TEMPLATE = "nvdcve-1.1-{}.json.gz"
+
+
+class Source(ABC):
+    @abstractmethod
+    def get_cve_data():
+        pass
