+++ cve_bin_tool/sources/curl.py
@@ -1,0 +1,147 @@
+from __future__ import annotations
+
+import glob
+import json
+import logging
+import os
+import re
+from typing import Any
+
+import aiohttp
+from bs4 import BeautifulSoup
+
+from cve_bin_tool.async_utils import FileIO, RateLimiter
+from cve_bin_tool.error_handler import (
+    CVEDataForCurlVersionNotInCache,
+    ErrorHandler,
+    ErrorMode,
+)
+from cve_bin_tool.log import LOGGER
+from cve_bin_tool.sources import DISK_LOCATION_BACKUP, DISK_LOCATION_DEFAULT, Source
+
+logging.basicConfig(level=logging.DEBUG)
+
+
+class Curl(Source):
+
+    CACHEDIR = DISK_LOCATION_DEFAULT
+    BACKUPCACHEDIR = DISK_LOCATION_BACKUP
+    LOGGER = LOGGER.getChild("CVEDB")
+    CURL_CVE_FILENAME_TEMPLATE = "curlcve-{}.json"
+
+    def __init__(self):
+        self.cve_list = None
+        self.cachedir = self.CACHEDIR
+        self.backup_cachedir = self.BACKUPCACHEDIR
+        self.error_mode = ErrorMode.TruncTrace
+        self.session = None
+        self.affected_data = None
+
+    async def get_cve_data(self):
+        await self.fetch_cves()
+        self.get_cve_list()
+
+        return None, self.affected_data
+
+    async def fetch_cves(self):
+        if not self.session:
+            connector = aiohttp.TCPConnector(limit_per_host=19)
+            self.session = RateLimiter(
+                aiohttp.ClientSession(connector=connector, trust_env=True)
+            )
+
+        versions = await self.get_curl_versions(self.session)
+
+        for version in versions:
+            await self.download_curl_version(self.session, version)
+
+        await self.session.close()
+
+    @staticmethod
+    async def get_curl_versions(session: RateLimiter) -> list[str]:
+        regex = re.compile(r"vuln-(\d+.\d+.\d+)\.html")
+        async with await session.get(
+            "https://curl.haxx.se/docs/vulnerabilities.html"
+        ) as response:
+            response.raise_for_status()
+            html = await response.text()
+        matches = regex.finditer(html)
+        return [match.group(1) for match in matches]
+
+    async def download_curl_version(self, session: RateLimiter, version: str) -> None:
+        async with await session.get(
+            f"https://curl.haxx.se/docs/vuln-{version}.html"
+        ) as response:
+            response.raise_for_status()
+            html = await response.text()
+        soup = BeautifulSoup(html, "html.parser")
+        table = soup.find("table")
+        if not table:
+            return
+        headers = table.find_all("th")
+        headers = list(map(lambda x: x.text.strip().lower(), headers))
+        self.LOGGER.debug(headers)
+        rows = table.find_all("tr")
+        json_data = []
+        for row in rows:
+            cols = row.find_all("td")
+            values = (ele.text.strip() for ele in cols)
+            data = dict(zip(headers, values))
+            if data:
+                json_data.append(data)
+        filepath = os.path.abspath(
+            os.path.join(self.cachedir, f"curlcve-{version}.json")
+        )
+        async with FileIO(filepath, "w") as f:
+            await f.write(json.dumps(json_data, indent=4))
+
+    def load_curl_version(self, version: str) -> dict[str, Any]:
+        """
+        Return the dict of CVE data for the given curl version.
+        """
+        filename = os.path.join(
+            self.cachedir, self.CURL_CVE_FILENAME_TEMPLATE.format(version)
+        )
+        # Check if file exists
+        if not os.path.isfile(filename):
+            with ErrorHandler(mode=self.error_mode, logger=self.LOGGER):
+                raise CVEDataForCurlVersionNotInCache(version)
+        # Open the file and load the JSON data, log the number of CVEs loaded
+        with open(filename, "rb") as fileobj:
+            cves_for_version = json.load(fileobj)
+            self.LOGGER.debug(
+                f"Curl Version {version} has {len(cves_for_version)} CVEs in dataset"
+            )
+            return cves_for_version
+
+    def curl_versions(self) -> list[str]:
+        """
+        Return the versions we have Curl data for.
+        """
+        regex = re.compile(r"curlcve-(\d+.\d+.\d).json")
+        return [
+            regex.search(filename).group(1)
+            for filename in glob.glob(os.path.join(self.cachedir, "curlcve-*.json"))
+        ]
+
+    def get_cve_list(self):
+        self.affected_data = []
+
+        for version in self.curl_versions():
+            cve_list = self.load_curl_version(version)
+
+            for cve in cve_list:
+                affected = {
+                    "cve_id": cve["cve"],
+                    "vendor": "haxx",
+                    "product": "curl",
+                    "version": version,
+                    "versionStartIncluding": cve["from version"],
+                    "versionStartExcluding": "",
+                    "versionEndIncluding": cve["to and including"],
+                    "versionEndExcluding": "",
+                }
+
+                self.affected_data.append(affected)
+
+        return self.affected_data
